+++ spec/cli/dummy/overwrite/transpile_me.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class A
+  def self.transform(val)
+    val => {status:}
+    val
+  end
+end